+++ src/main.rs
@@ -3,6 +3,7 @@ mod engine;
 mod icecast;
 mod listen;
 mod module;
+mod mux;
 mod rtmp;
 mod source;
 mod throttle;
